+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Increase limit for top delegates from 5 to 100
+
 ## [0.4.15] - 2025-04-25
 
 ### Changed