+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Increase limit for top delegates from 5 to 100
+- Fix status response for non existing proposal id
+- Extend proposals get by filters endpoint for getting short proposal info 
+
 ## [0.4.16] - 2025-05-27
 
 ### Added