+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.17] - 2025-05-27
+
 ### Changed
 - Increase limit for top delegates from 5 to 100
 - Fix status response for non existing proposal id